requirements.txt
langchain
langchain-community
langchain-groq
langchin-chroma
python-dotenv
pypdf
sentence_transformers
langchain-huggingface
faiss-cpu
streamlit
streamlit-extras
